fix(dep-resolution): correctly report resolution method when resolving dynamically (semgrep/semgrep-proprietary#2667)
We previously were incorrectly setting the resolution method to lockfile
parsing
even when dependencies were resolved another way. This was not used
anywhere, so
there are no consequences of this bug. This PR changes the way that we
set
resolution method so that it reflects the actual resolution method.

synced from Pro 36cc4745f562976642e7164af72a1de8548d943d