Add tests for pysemgrep bugs (semgrep/semgrep-proprietary#2686)
I discovered that the path matching function `fnmatch` that we've been
using all along is not suitable for matching whole paths but only for
matching individual path segments. It doesn't support `**` since it
doesn't recognize `/` as a special character.

It's surprising that we've been able to go that far without noticing it.
I left a comment with the following possible solutions:
1. Do nothing and migrate to osemgrep
2. Upgrade to Python 3.13 and use `PurePath.full_match`
3. Install the wcmatch library and use
`wcmatch.pathlib.PurePath(str).globmatch(pat)`

This branch adds tests for a few problematic paths with
`.semgrepignore`, and tests for `--include`, and `--exclude`. These
tests pass with osemgrep and are marked as failing with pysemgrep (as
comments).

synced from Pro 09aeaa67d46cc5e1447949707e3a74d372b2118f