chore: relevant info in for otel metrics (semgrep/semgrep-proprietary#2633)
In order to get opentelemetry metrics in an organized way, we want to be
able to group metrics by the scan they happen in, so we can see metrics
by either scan, or globally across all recent scans. the latter is
already possible, but the former is not, as there are no attributes
indicating which scan a metric came from.

This PR reorganizes/cleans up some code related to attributes, and then
uses SEMGREP_JOB_URL as the value for `service.instance.id`, which is an
attribute that is automatically applied to all metrics sent from otel.
This now allows us to group metrics by scan (or `instance`), in grafana
(see test plan. for an example of how that looks).

Semantically `service.instance.id` doesn't have much meaning to
opentelemetry
([ref](https://opentelemetry.io/docs/specs/semconv/attributes-registry/service/#service-attributes)),
but prometheus (which opentelemetry exports to, and is the source of
data for grafana), sees it as something more specific. Prometheus
expects instance to be a `host:port`
([ref](https://prometheus.io/docs/concepts/jobs_instances/#jobs-and-instances)).
Because of this I leaned towards using SEMGREP_JOB_URL. The main
downside here being this is an environment variable, so it won't be set
for scans where users haven't explicitly set this environment variable,
and in turn those metrics won't be grouped by scan.

Another option instead of using SEMGREP_JOB_URL, would be to use the
scan ID. the downside here is that if we want to add a nice button to
argo workflows to link to the metrics of a scan, the workflow would need
to know the scan ID ahead of time. Curious to hear what the reviewer
thinks makes sense for the instance id.

## Test plan
run semgrep
``` bash
SEMGREP_JOB_URL="https://argoworkflows-prod.corp.r2c.dev/workflows/console/zcs-agent-foobar" WORKFLOW_NAMESPACE="console" semgrep scan --config p/auto --trace --pro
```
Metrics grouped by instance in grafana!!! See
[here](https://g-8669019a19.grafana-workspace.us-west-2.amazonaws.com/explore?schemaVersion=1&panes=%7B%22mtk%22:%7B%22datasource%22:%22IYx4ui2Iz%22,%22queries%22:%5B%7B%22refId%22:%22A%22,%22expr%22:%22process_runtime_ocaml_gc_minor_allocated_B_total%7Binstance%3D%5C%22https:%2F%2Fargoworkflows-prod.corp.r2c.dev%2Fworkflows%2Fconsole%2Fzcs-agent-foobar%5C%22%7D%22,%22range%22:true,%22instant%22:true,%22datasource%22:%7B%22type%22:%22prometheus%22,%22uid%22:%22IYx4ui2Iz%22%7D,%22editorMode%22:%22builder%22,%22legendFormat%22:%22__auto%22,%22useBackend%22:false,%22disableTextWrap%22:false,%22fullMetaSearch%22:false,%22includeNullMetadata%22:true%7D%5D,%22range%22:%7B%22from%22:%22now-1h%22,%22to%22:%22now%22%7D%7D%7D&orgId=1)

synced from Pro 876f5caeccafcc61f91518b69963e49df18bda74