Fix semgrep failures on symlinks to paths without read permissions (semgrep/semgrep-proprietary#2798)
This PR fixes problems where semgrep was not handling certain broken
symlinks correctly. In particular, a symlink to a path such as
`/root/foo` where the current user doesn't have read access to `/root`
was resulting in an exception both in pysemgrep and in osemgrep. This
had to be fixed in multiple spots in Python and in OCaml where syscalls
to `stat` (which raises an exception) was being called before checking
that we had read access to the file.

Testing for missing file permissions is difficult, as shown in an
earlier attempt. The tests we have in place include:
* `test_missing_file`: it doesn't test in particular for missing
permissions but triggers the same errors
* `test_target_selection`: runs `--x-ls` to list the files of our test
repo https://github.com/semgrep/semgrep-test-project1 where the file
`linux/link-to-file-in-dir-without-read-perm` is a broken symlink. As a
target file that is not a scanning root, it will be reported as a
skipped target.

Manual testing for `semgrep ci` on a test repo containing a broken
symlink revealed extra errors that needed fixing on the Python side.

synced from Pro 693acc4d6ac1dea9306a355cf418b11ca0a08644