Enable reporting of unused variables with flake8 (semgrep/semgrep-proprietary#2963)
@0xDC0DE the unused-variable checker found that `configs_with_fixtests`
was unused. This is some code that you introduced. I didn't bother
trying to make it work. Fixing the misspelling to `config_with_fixtests`
results in a type error reported by mypy so I commented it out. If you
remember what this is supposed to do or how it should be tested, please
chime in.

synced from Pro 8993392f7757f6d45511d84c450628c595029cc2