Remove perf/perf-matching/ (semgrep/semgrep-proprietary#2998)
This was an old benchmark we ran to compare different matching
strategy. Most matching-specific optimizations got removed because -fast
was
usually better.
I doubt anyone in the last year looked at the results of those
benchs at https://dashboard.semgrep.dev/benchmarks
and there is no cron I'm aware of alerting if something got bad,
so let's just remove the cruft.

The reason I discovered this dead code is because in my PR
https://github.com/semgrep/semgrep-proprietary/pull/2985
I changed the behavior of semgrep-core -lang, and this causes
the OSS release to fail because the perf-matching script was using
the -lang regex which was not supported at all.
At least we should move this CI check from OSS to Pro so that
we would detect the regression earlier. Then we should
just modify the script to not use semgrep-core but osemgrep which
does support the flexible -lang regex (which really should now
be called -analyzer regex).
But just simpler to remove the script and everything related
to perf-matching.

test plan:
CI checks
git grep perf-matching
is now empty

synced from Pro def33863bd7dbd88b741a93614cd7447fba205df