feat: expose metavariable-name.modules (semgrep/semgrep-proprietary#3188)
Previously we only allowed metavariable-name to appear with a
`semgrep-internal-` prefix. This allows for it to appear unqualified as
long as only the `module` constraint is used (i.e., `kind` is not used).
Note also that this heavily relies on Pro naming, and thus still
requires the Pro engine.

synced from Pro 1dd652823fcc0812a428e0524d7bc2f60b15a35f