Test TargetManager with use_semgrepignore_v2=True (semgrep/semgrep-proprietary#3099)
This adds the `--semgrepignore-v2` option to the CLI but it's off by
default. The tests that invoke `TargetManager()` directly now run with
this setting both on and off.

The end-to-end tests use the default setting. Many of them fail when I
turn it on.

synced from Pro 679020d1df482d229c6e1d04a95be2db21f2d60e