feat: simplify semgrep_interfaces setup (semgrep/semgrep-proprietary#3273)
Right now we aren't installing any of the deps needed to build/test semgrep-interfaces, which is needed for the core-e2e-test target. This PR changes that so we install the needed deps to run those tests (check-jsonschema). It also installs the deps needed to generate the interfaces in the first place, which means in the future we could easily make building the interfaces a dune rule.

Please approve https://github.com/semgrep/semgrep-interfaces/pull/361 first so we can set the submodule ref to a develop commit.

## Test plan
check-semgrep-oss passes

synced from Pro b554c51ff7b354b6040379bfee051169c890f0de