AST_to_IL: JS/TS: Translate `for (let x of E)` as a for-each (semgrep/semgrep-proprietary#3468)
This allows us to remove this built-in custom propagator:

    for ($TO of $FROM) {...}

which has been causing perf issues in combination with #3043.

Running `semgrep -c p/default --pro HashBrownCMS/hashbrown-cms` became
like 10x slower.

We still need to investigate the reason for that perf regression,
because in principle it should be OK having that propagator (despite in
this case we should have fixed the AST-to-IL translation in the first
place). I think it is probably related to being part of a loop and the
implicit clean-up of the LHS in an assigment where the RHS is not
tainted... This was causing a fixpoint timeout on every single `for (let
x of E)` loop.

Fixes: 12a0d42c3b1 ("tainting: Make out-of-order propagation happen instantly (semgrep/semgrep-proprietary#3043)")

test plan:
make test
semgrep -c p/default --pro HashBrownCMS/hashbrown-cms

synced from Pro 3faa997b3ea463f528543bddeb17b34fc736b526